+++ app/src/main/java/org/sopt/certi/presentation/ui/resume/component/ResumeDescriptionSection.kt
@@ -61,7 +61,7 @@ fun ResumeDescriptionSection(
 
 @Preview(showBackground = true)
 @Composable
-fun ResumeDescriptionSectionPreview() {
+private fun ResumeDescriptionSectionPreview() {
     CERTITheme {
         ResumeDescriptionSection(
             resumeListItem = ResumeListData(
